made search friend thunk into its own file
