+++ client/store/friends-list-reducer/index.js
@@ -3,12 +3,14 @@ import nearbyUsers from './nearby-users';
 import sentRequests from './sent-requests';
 import receivedRequests from './received-requests';
 import friends from './friends';
+import userList from './find-friends'
 
 export const reducer = combineReducers({
   nearbyUsers,
   sentRequests,
   receivedRequests,
   friends,
+  userList
 });
 
 export default reducer;
@@ -17,3 +19,4 @@ export * from './nearby-users';
 export * from './sent-requests';
 export * from './received-requests';
 export * from './friends';
+export * from './find-friends'
